chore: begin migrating to typescript
