--- config/dbConecction.js
@@ -1,13 +1,0 @@
-const mongoose = require("mongoose");
-
-const connectDb = async () => {
-  try {
-    const connect = await mongoose.connect(process.env.CONNECTION_STRING);
-    console.log("Database connected: ", connect.connection.host, connect.connection.name)
-  } catch(error) {
-    console.log(error);
-    process.exit(1);
-  };
-};
-
-module.exports = connectDb;
