+++ src/config/dbConnection.ts
@@ -1,0 +1,13 @@
+import mongoose from "mongoose";
+
+const connectDb = async (): Promise<void> => {
+  try {
+    const connect = await mongoose.connect(process.env.CONNECTION_STRING as string);
+    console.log("Database connected: ", connect.connection.host, connect.connection.name);
+  } catch (error) {
+    console.log(error);
+    process.exit(1);
+  }
+};
+
+export default connectDb;
